Rename Python module to scippy.
