+++ scippy/.gitignore
