+++ scippy/compat/test/mantid_test.py
