+++ scippy/demo/img/1280px-NumPy_logo.png
