+++ scippy/demo/img/WorkspaceHierarchy.jpg
