+++ scippy/demo/img/cumulative-algorithm-count.png
