+++ scippy/demo/img/dataset-3d-two-data-variables.png
