+++ scippy/demo/img/dataset-diagram-logo.png
