+++ scippy/demo/img/google-trends.png
