+++ scippy/examples/3d_volume_handling.py
