+++ scippy/examples/3d_volume_xarray_plot.py
