+++ scippy/examples/create_3d_volume_plotting_slices.py
