+++ scippy/examples/dataset_as_table.py
