+++ scippy/examples/demo-part1.py
