+++ scippy/examples/demo-part2.py
