+++ scippy/examples/demo-part3.py
