+++ scippy/examples/workspace2d_treatment.py
