+++ scippy/test/test_datasetslice.py
@@ -1,6 +1,6 @@
 import unittest
 
-from dataset import *
+from scippy import *
 import numpy as np
 import operator
 
@@ -15,11 +15,11 @@ class TestDatasetSlice(unittest.TestCase):
 
     def test_type(self):
         ds_slice = self._d.subset["a"]
-        self.assertEqual(type(ds_slice), dataset.DatasetSlice)
+        self.assertEqual(type(ds_slice), scippy.DatasetSlice)
 
     def test_extract_slice(self):
         ds_slice = self._d.subset["a"]
-        self.assertEqual(type(ds_slice), dataset.DatasetSlice)
+        self.assertEqual(type(ds_slice), scippy.DatasetSlice)
         # We should have just one data variable
         self.assertEqual(1, len([var for var in ds_slice if var.is_data]))
         # We should have just one coord variable
