+++ scippy/test/test_variable.py
@@ -1,6 +1,6 @@
 import unittest
 
-from dataset import *
+from scippy import *
 import numpy as np
 
 def make_variables():
